feat(project): create ProjectEntity with validation and properties for project management fix(imports): update import paths for ProjectEntity to reflect new structure refactor(typeorm): change entity path in TypeORM configuration to use entities folder style(controller): improve summary in createCrudController for better readability
