+++ server/src/scaffold/controllers/scaffold-crud.controller.ts
@@ -8,6 +8,10 @@ import {scaffoldSort} from '@/scaffold/utils/scaffold-sort';
 import {Get, Query, Type} from '@nestjs/common';
 import {ApiOkResponse, ApiOperation} from '@nestjs/swagger';
 
+function toHuman<Entity extends ScaffoldEntity>(entity: Type<Entity>) {
+    return entity.name.replace(/Entity$/, '');
+}
+
 export function createCrudController<Entity extends ScaffoldEntity>(
     entity: Type<Entity>
 ) {
@@ -20,7 +24,7 @@ export function createCrudController<Entity extends ScaffoldEntity>(
 
         @Get()
         @ApiOperation({
-            summary: `List all ${entity.name}`
+            summary: `List all ${toHuman(entity)}`
         })
         @DtoResponse([entity])
         @ApiOkResponse({
