+++ server/src/typeorm.config.ts
@@ -7,7 +7,7 @@ dotenv.config();
 const dataSource = new DataSource({
     type: 'sqlite',
     database: process.env.PROJECTS_FOLDER + '/aigraphr.sqlite',
-    entities: [`${__dirname}/models/*.entity{.ts,.js}`],
+    entities: [`${__dirname}/entities/*.entity{.ts,.js}`],
     subscribers: [`${__dirname}/subscribers/*.subscriber{.ts,.js}`],
     migrationsRun: false,
     migrations: [`${__dirname}/migrations/*{.ts,.js}`]
